+++ front/components/pages/disks/index.tsx
@@ -2,6 +2,10 @@ import {Table} from 'react-bootstrap';
 import TableRow from '~/components/table/row';
 import TableRows from '~/components/table/rows';
 
+interface Props {
+  data: any
+}
+
 const getPartitions = data => {
   const partitions = [];
   data.partitions.forEach((partition, index) => {
@@ -20,7 +24,7 @@ const getPartitions = data => {
   return partitions;
 };
 
-const Disks = props => {
+const Disks = (props: Props) => {
   const {data} = props;
   if (!Object.keys(data).length) return null;
