--- front/components/pages/index.js
@@ -1,8 +1,0 @@
-export * from './cpu';
-export * from './memory';
-export * from './swap';
-export * from './disks';
-export * from './network';
-export * from './networkConnections';
-export * from './processes';
-export * from './devices';
