+++ front/components/pages/swap/index.tsx
@@ -1,17 +1,21 @@
 import {Table} from 'react-bootstrap';
 import TableRow from '~/components/table/row';
 
-const Memory = props => {
+interface Props {
+  data: any
+}
+
+const Swap = (props: Props) => {
   const {data} = props;
   if (!Object.keys(data).length) return null;
 
   return <Table striped bordered hover size="sm">
     <tbody>
       <TableRow head={'容量'} body={`${data.total} GB`} />
-      <TableRow head={'空き容量'} body={`${data.available} GB`} />
+      <TableRow head={'空き容量'} body={`${data.free} GB`} />
       <TableRow head={'使用量'} body={`${data.used} GB`} />
     </tbody>
   </Table>
 };
 
-export {Memory};
+export {Swap};
