+++ front/next.config.js
@@ -2,10 +2,6 @@ const withCSS = require('@zeit/next-css');
 const path = require('path');
 
 module.exports = withCSS({
-  /* config options here */
-});
-
-module.exports = {
   devIndicators: {
     autoPrerender: false,
   },
@@ -16,4 +12,4 @@ module.exports = {
   env: {
     apiUrl: 'http://localhost:5000/api'
   }
-};
+});
