+++ front/package.json
@@ -15,5 +15,12 @@
     "react": "16.13.0",
     "react-bootstrap": "^1.0.0-beta.17",
     "react-dom": "16.13.0"
+  },
+  "devDependencies": {
+    "@types/axios": "^0.14.0",
+    "@types/bootstrap": "^4.3.2",
+    "@types/node": "^13.11.1",
+    "@types/react-bootstrap": "^0.32.20",
+    "typescript": "^3.8.3"
   }
 }
