+++ front/pages/memory/index.tsx
@@ -1,17 +1,17 @@
 import {useState, useEffect} from 'react';
-import {Swap as SwapComponent} from '~/components/pages/swap';
+import {Memory as MemoryComponent} from '~/components/pages/memory';
 import Layout from '~/components/layout';
-import {getSwapApi} from '~/api';
+import {getMemoryApi} from '~/api';
 
-const Swap = () => {
-  const [isLoading, setIsLoading] = useState(false);
+const Memory = () => {
+  const [isLoading, setIsLoading] = useState<boolean>(false);
   const [data, setData] = useState({});
 
   useEffect(() => {
     (async () => {
       setIsLoading(true);
       try {
-        const response = await getSwapApi();
+        const response = await getMemoryApi();
         setData(response.data);
       } catch(error) {
         console.log(error.response);
@@ -21,14 +21,14 @@ const Swap = () => {
     })();
   }, []);
 
-  return <Layout activeKey={'swap'}>
-    <h3>Swap</h3>
+  return <Layout activeKey={'memory'}>
+    <h3>Memory</h3>
     {
       isLoading
         ? <p className='text-center'>Loading...</p>
-        : <SwapComponent data={data} />
+        : <MemoryComponent data={data} />
     }
   </Layout>
 }
 
-export default Swap;
+export default Memory;
