+++ front/pages/network/index.tsx
@@ -4,7 +4,7 @@ import Layout from '~/components/layout';
 import {getNetworkApi} from '~/api';
 
 const Network = () => {
-  const [isLoading, setIsLoading] = useState(false);
+  const [isLoading, setIsLoading] = useState<boolean>(false);
   const [data, setData] = useState({});
 
   useEffect(() => {
